Pop LightsBridge into separate component
